+++ js/src/admin/index.ts
@@ -5,10 +5,10 @@ app.initializers.add("nearata-minecraft-avatars", (app) => {
         setting: "nearata-minecraft-avatars.retrieve_avatar_on_register",
         type: "boolean",
         label: app.translator.trans(
-            "nearata-minecraft-avatars.admin.retrieve_avatar_on_registering.label"
+            "nearata-minecraft-avatars.admin.retrieve_avatar_on_register.label"
         ),
         help: app.translator.trans(
-            "nearata-minecraft-avatars.admin.retrieve_avatar_on_registering.help"
+            "nearata-minecraft-avatars.admin.retrieve_avatar_on_register.help"
         ),
     });
 });